data written to files!

+++ BLE_Benchmark/gatt-client/src/main/java/edu/nd/cse/gatt_client/BenchmarkProfileClientCallback.java
@@ -13,7 +13,7 @@ public interface BenchmarkProfileClientCallback {
 
     public void onBytesSentAvailable (long bytesSent);
 
-    public void onStartupLatencyAvailable (float startLatency);
+    public void onStartupLatencyAvailable (long startLatency);
 
     public void onThroughputAvailable (float throughput);
